+++ src/app/indicator-sharing/indicator-sharing-filters/indicator-sharing-filters.component.scss
@@ -1,0 +1,3 @@
+#indicatorSharingFilters {
+    padding: 24px;
+}
