+++ src/app/indicator-sharing/indicator-sharing-list/indicator-sharing-list.component.scss
@@ -1,12 +1,32 @@
 @import '../../../styles/_variables.scss';
 
-#filterWrapper {
-    padding-right: 0;
-    padding-top: 80px;
+#filterTitle {
+  height: 64px;
+  background: $analytic-hub-primary-lighter;
+  padding-left: 7px;
+}
+
+#mainWindow {
+  min-height: 90vh;
+}
+
+#filterContainer {
+  width: 360px;
 }
 
 #searchWrapper {
   background-color: $analytic-hub-primary-light;
   height: 128px;
+  transition: height 0.2s ease-in-out;
+}
+
+.filterOpen {
+  height: 64px !important;
+}
+
+#showFiltersButton {
+  position: absolute;
+  top: 13px;
+  left: 10px;
 }
