+++ src/app/indicator-sharing/indicator-sharing-list/indicator-sharing-list.component.ts
@@ -11,11 +11,13 @@ import * as fromIndicatorSharing from '../store/indicator-sharing.reducers';
 import * as indicatorSharingActions from '../store/indicator-sharing.actions';
 import { Constance } from '../../utils/constance';
 import { IndicatorBase } from '../models/indicator-base-class';
+import { fadeInOut } from '../../global/animations/fade-in-out';
 
 @Component({
     selector: 'indicator-sharing-list',
     templateUrl: 'indicator-sharing-list.component.html',
     styleUrls: ['indicator-sharing-list.component.scss'],
+    animations: [fadeInOut],
     changeDetection: ChangeDetectionStrategy.OnPush
 })
 
@@ -25,6 +27,7 @@ export class IndicatorSharingListComponent extends IndicatorBase implements OnIn
     public filteredIndicators: any[];
     public DEFAULT_LENGTH: number = 10;
     public searchParameters;
+    public filterOpen: boolean = false;
 
     constructor(
         private indicatorSharingService: IndicatorSharingService, 
@@ -175,5 +178,13 @@ export class IndicatorSharingListComponent extends IndicatorBase implements OnIn
         } else {
             return this.displayedIndicators.length < this.filteredIndicators.length;
         }
+    }
+
+    public openedStart() {
+        this.filterOpen = true;
+    }
+
+    public closedStart() {
+        this.filterOpen = false;
     }
 }
