+++ src/app/indicator-sharing/indicator-sharing.module.ts
@@ -3,7 +3,7 @@ import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
 import { FormsModule, ReactiveFormsModule } from '@angular/forms';
 import { MatSelectModule, MatChipsModule, MatTooltipModule, MatCardModule,
-    MatTabsModule, MatInputModule, MatIconModule, MatButtonModule, MatDatepickerModule, MatNativeDateModule, MatRadioModule, MatListModule, MatStepperModule, MatDialogModule } from '@angular/material';
+    MatTabsModule, MatInputModule, MatIconModule, MatButtonModule, MatDatepickerModule, MatNativeDateModule, MatRadioModule, MatListModule, MatStepperModule, MatDialogModule, MatSidenavModule } from '@angular/material';
 import { StoreModule } from '@ngrx/store';
 import { EffectsModule } from '@ngrx/effects';
 
@@ -22,11 +22,11 @@ import { IndicatorSharingEffects } from './store/indicator-sharing.effects';
 import { IndicatorSharingLayoutComponent } from './indicator-sharing-layout/indicator-sharing-layout.component';
 import { IndicatorSharingListComponent } from './indicator-sharing-list/indicator-sharing-list.component';
 import { IndicatorCardComponent } from './indicator-card/indicator-card.component';
-import { IndicatorSharingSearchComponent } from './indicator-sharing-search/indicator-sharing-search.component';
 import { AddIndicatorComponent } from './add-indicator/add-indicator.component';
 import { IndicatorDetailsComponent } from './indicator-details/indicator-details.component';
 import { SearchBarComponent } from './search-bar/search-bar.component';
 import { IndicatorSharingSortComponent } from './indicator-sharing-sort/indicator-sharing-sort.component';
+import { IndicatorSharingFiltersComponent } from './indicator-sharing-filters/indicator-sharing-filters.component';
 
 const matModules = [
     MatButtonModule,
@@ -43,6 +43,7 @@ const matModules = [
     MatStepperModule,
     MatTabsModule,
     MatTooltipModule,
+    MatSidenavModule
 ];
 
 @NgModule({
@@ -64,10 +65,10 @@ const matModules = [
         IndicatorSharingListComponent,
         IndicatorCardComponent,
         AddIndicatorComponent,
-        IndicatorSharingSearchComponent,
         IndicatorDetailsComponent,
         SearchBarComponent,
-        IndicatorSharingSortComponent
+        IndicatorSharingSortComponent,
+        IndicatorSharingFiltersComponent
     ],
     providers: [
         IndicatorSharingService
